Cleanup and bug found.
Parts of the code have been cleaned up and in the process, 2 bugs were found. Firstly, we were naively interpolating for a while now because most scenes showed no difference. Secondly, the correct interpolation fails when triangles a coplanar to the x or y plane. This is an easy fix but will require some time to complete. It is important to note that technically any triangle not coplanar to the z plane is "wrong" but the math works out in such a way that the answer is identical anyway so long as both x and y values are distinct.