Created new interpolator class that handles edge cases.
The new interpolator class is meant to replace the old struct. It also (automatically) interpolates in world space, thus, does depth corrected interpolation at the same time. It can also handle cases where the triangle is perpendicular to the z plane by reorienting the triangle such that it is parallel to the z plane. I'm sure there is a better and more efficient way of doing this but for now this will have to suffice.